pyproject.toml
[tool.poetry]
name = "project-prompt"
version = "1.0.0"
description = "Asistente inteligente para análisis y documentación de proyectos usando IA"
authors = ["ProjectPrompt Team <info@project-prompt.org>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/project-prompt/project-prompt"
documentation = "https://docs.project-prompt.org"
keywords = ["ai", "prompt-engineering", "documentation", "project-analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.9.0"
rich = "^13.5.0"
openai = "^1.0.0"
anthropic = "^0.5.0"
pyyaml = "^6.0.1"
python-frontmatter = "^1.0.0"
jinja2 = "^3.0.0"
tabulate = "^0.9.0"
requests = "^2.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
mypy = "^1.3.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
project-prompt = "src.main:app"

[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/project-prompt/project-prompt/issues"
"Documentation" = "https://docs.project-prompt.org"
"Source Code" = "https://github.com/project-prompt/project-prompt"
